fix: rename helpers file, move TracePrintListen into its own
